Fix the edit model in inventory

+++ database/migrations/2026_08_02_083720_create_inventories_table.php
@@ -22,9 +22,9 @@ return new class extends Migration
             $table->unsignedBigInteger('supplier_id')
                 ->nullable();
 
-            $table->integer('quantity');
+            $table->decimal('quantity', 10, 2);
 
-            $table->integer('remaining_quantity');
+            $table->decimal('remaining_quantity', 10, 2);
 
             $table->string('batch_number')->unique();
